refactor : move fireFunctionMultipleTimes function from App component to utils module
